Refactor agent sidebar
