+++ web_src/src/components/AgentSidebar/AiBuilderInputForm.tsx
