+++ web_src/src/ui/BuildingBlocksSidebar/index.tsx
@@ -12,7 +12,7 @@ import { ComponentBase } from "../componentBase";
 import { CategorySection } from "./CategorySection";
 import type { BuildingBlock, BuildingBlockCategory } from "./types";
 
-export type { AgentContext, AgentMode } from "./agentChat";
+export type { AgentContext, AgentMode } from "@/components/AgentSidebar/agentChat";
 export type { BuildingBlock, BuildingBlockCategory } from "./types";
 
 export interface BuildingBlocksSidebarProps {
